+++ codegen/impl/src/es/event/revised.rs
@@ -1,4 +1,4 @@
-//! `#[derive(event::Versioned)]` macro implementation.
+//! `#[derive(event::Revised)]` macro implementation.
 
 use std::num::NonZeroU16;
 
@@ -7,7 +7,7 @@ use quote::quote;
 use syn::spanned::Spanned as _;
 use synthez::{ParseAttrs, Required, ToTokens};
 
-/// Expands `#[derive(event::Versioned)]` macro.
+/// Expands `#[derive(event::Revised)]` macro.
 ///
 /// # Errors
 ///
@@ -20,20 +20,20 @@ pub fn derive(input: TokenStream) -> syn::Result<TokenStream> {
     Ok(quote! { #definition })
 }
 
-/// Helper attributes of `#[derive(event::Versioned)]` macro.
+/// Helper attributes of `#[derive(event::Revised)]` macro.
 #[derive(Debug, Default, ParseAttrs)]
 pub struct Attrs {
-    /// Value of [`event::Versioned::NAME`][0] constant.
+    /// Value of [`event::Revised::NAME`][0] constant.
     ///
-    /// [0]: arcane_core::es::event::Versioned::NAME
+    /// [0]: arcane_core::es::event::Revised::NAME
     #[parse(value)]
     pub name: Required<syn::LitStr>,
 
-    /// Value of [`event::Versioned::VERSION`][0] constant.
+    /// Value of [`event::Revised::REVISION`][0] constant.
     ///
-    /// [0]: arcane_core::es::event::Versioned::VERSION
-    #[parse(value, alias = ver, validate = can_parse_as_non_zero_u16)]
-    pub version: Required<syn::LitInt>,
+    /// [0]: arcane_core::es::event::Revised::REVISION
+    #[parse(value, alias = rev, validate = can_parse_as_non_zero_u16)]
+    pub revision: Required<syn::LitInt>,
 }
 
 /// Checks whether the given `value` can be parsed as [`NonZeroU16`].
@@ -41,12 +41,12 @@ fn can_parse_as_non_zero_u16(value: &Required<syn::LitInt>) -> syn::Result<()> {
     syn::LitInt::base10_parse::<NonZeroU16>(value).map(drop)
 }
 
-/// Representation of a struct implementing [`event::Versioned`][0], used for
+/// Representation of a struct implementing [`event::Revised`][0], used for
 /// code generation.
 ///
-/// [0]: arcane_core::es::event::Versioned
+/// [0]: arcane_core::es::event::Revised
 #[derive(Debug, ToTokens)]
-#[to_tokens(append(impl_event_versioned, gen_uniqueness_glue_code))]
+#[to_tokens(append(impl_event_revised, gen_uniqueness_glue_code))]
 pub struct Definition {
     /// [`syn::Ident`](struct@syn::Ident) of this structure's type.
     pub ident: syn::Ident,
@@ -54,16 +54,16 @@ pub struct Definition {
     /// [`syn::Generics`] of this structure's type.
     pub generics: syn::Generics,
 
-    /// Value of [`event::Versioned::NAME`][0] constant in the generated code.
+    /// Value of [`event::Revised::NAME`][0] constant in the generated code.
     ///
-    /// [0]: arcane_core::es::event::Versioned::NAME
+    /// [0]: arcane_core::es::event::Revised::NAME
     pub event_name: syn::LitStr,
 
-    /// Value of [`event::Versioned::VERSION`][0] constant in the generated
+    /// Value of [`event::Revised::REVISION`][0] constant in the generated
     /// code.
     ///
-    /// [0]: arcane_core::es::event::Versioned::VERSION
-    pub event_version: syn::LitInt,
+    /// [0]: arcane_core::es::event::Revised::REVISION
+    pub event_revision: syn::LitInt,
 }
 
 impl TryFrom<syn::DeriveInput> for Definition {
@@ -84,42 +84,42 @@ impl TryFrom<syn::DeriveInput> for Definition {
             ident: input.ident,
             generics: input.generics,
             event_name: attrs.name.into_inner(),
-            event_version: attrs.version.into_inner(),
+            event_revision: attrs.revision.into_inner(),
         })
     }
 }
 
 impl Definition {
-    /// Generates code to derive [`event::Versioned`][0] trait.
+    /// Generates code to derive [`event::Revised`][0] trait.
     ///
-    /// [0]: arcane_core::es::event::Versioned
+    /// [0]: arcane_core::es::event::Revised
     #[must_use]
-    pub fn impl_event_versioned(&self) -> TokenStream {
+    pub fn impl_event_revised(&self) -> TokenStream {
         let ty = &self.ident;
         let (impl_gens, ty_gens, where_clause) = self.generics.split_for_impl();
 
-        let (event_name, event_ver) = (&self.event_name, &self.event_version);
+        let (event_name, event_rev) = (&self.event_name, &self.event_revision);
 
         quote! {
             #[automatically_derived]
-            impl #impl_gens ::arcane::es::event::Versioned for #ty #ty_gens
+            impl #impl_gens ::arcane::es::event::Revised for #ty #ty_gens
                  #where_clause
             {
                 const NAME: ::arcane::es::event::Name = #event_name;
 
                 // SAFETY: Safe, as checked by proc macro in compile time.
-                const VERSION: ::arcane::es::event::Version = unsafe {
-                    ::arcane::es::event::Version::new_unchecked(#event_ver)
+                const REVISION: ::arcane::es::event::Revision = unsafe {
+                    ::arcane::es::event::Revision::new_unchecked(#event_rev)
                 };
             }
         }
     }
 
     /// Generates hidden machinery code used to statically check uniqueness of
-    /// [`Event::name`] and [`Event::version`].
+    /// [`Event::name`] and [`Event::revision`].
     ///
     /// [`Event::name`]: arcane_core::es::Event::name
-    /// [`Event::version`]: arcane_core::es::Event::version
+    /// [`Event::revision`]: arcane_core::es::Event::revision
     #[must_use]
     pub fn gen_uniqueness_glue_code(&self) -> TokenStream {
         let ty = &self.ident;
@@ -131,7 +131,7 @@ impl Definition {
         quote! {
             #[automatically_derived]
             #[doc(hidden)]
-            impl #impl_gens ::arcane::es::event::codegen::Versioned for
+            impl #impl_gens ::arcane::es::event::codegen::Revised for
                  #ty #ty_gens #where_clause
             {
                 #[doc(hidden)]
@@ -154,8 +154,8 @@ impl Definition {
                             "_",
                             ::std::column!(),
                         ),
-                        <Self as ::arcane::es::event::Versioned>::NAME,
-                        <Self as ::arcane::es::event::Versioned>::VERSION.get(),
+                        <Self as ::arcane::es::event::Revised>::NAME,
+                        <Self as ::arcane::es::event::Revised>::REVISION.get()
                     )]
                 }
             }
@@ -171,24 +171,24 @@ mod spec {
     #[test]
     fn derives_struct_impl() {
         let input = parse_quote! {
-            #[event(name = "event", version = 1)]
+            #[event(name = "event", revision = 1)]
             struct Event;
         };
 
         let output = quote! {
             #[automatically_derived]
-            impl ::arcane::es::event::Versioned for Event {
+            impl ::arcane::es::event::Revised for Event {
                 const NAME: ::arcane::es::event::Name = "event";
 
                 // SAFETY: Safe, as checked by proc macro in compile time.
-                const VERSION: ::arcane::es::event::Version = unsafe {
-                    ::arcane::es::event::Version::new_unchecked(1)
+                const REVISION: ::arcane::es::event::Revision = unsafe {
+                    ::arcane::es::event::Revision::new_unchecked(1)
                 };
             }
 
             #[automatically_derived]
             #[doc(hidden)]
-            impl ::arcane::es::event::codegen::Versioned for Event {
+            impl ::arcane::es::event::codegen::Revised for Event {
                 #[doc(hidden)]
                 const COUNT: usize = 1;
             }
@@ -209,8 +209,8 @@ mod spec {
                             "_",
                             ::std::column!(),
                         ),
-                        <Self as ::arcane::es::event::Versioned>::NAME,
-                        <Self as ::arcane::es::event::Versioned>::VERSION.get(),
+                        <Self as ::arcane::es::event::Revised>::NAME,
+                        <Self as ::arcane::es::event::Revised>::REVISION.get()
                     )]
                 }
             }
@@ -225,7 +225,7 @@ mod spec {
     #[test]
     fn name_arg_is_required() {
         let input = parse_quote! {
-            #[event(ver = 1)]
+            #[event(rev = 1)]
             struct Event;
         };
 
@@ -239,7 +239,7 @@ mod spec {
     }
 
     #[test]
-    fn version_arg_is_required() {
+    fn revision_arg_is_required() {
         let input = parse_quote! {
             #[event(name = "event")]
             struct Event;
@@ -249,15 +249,15 @@ mod spec {
 
         assert_eq!(
             err.to_string(),
-            "either `ver` or `version` argument of `#[event]` attribute is \
+            "either `rev` or `revision` argument of `#[event]` attribute is \
              expected to be present, but is absent",
         );
     }
 
     #[test]
-    fn errors_on_negative_version() {
+    fn errors_on_negative_revision() {
         let input = parse_quote! {
-            #[event(name = "event", ver = -1)]
+            #[event(name = "event", rev = -1)]
             struct Event;
         };
 
@@ -267,9 +267,9 @@ mod spec {
     }
 
     #[test]
-    fn errors_on_zero_version() {
+    fn errors_on_zero_revision() {
         let input = parse_quote! {
-            #[event(name = "event", version = 0)]
+            #[event(name = "event", revision = 0)]
             struct Event;
         };
 
@@ -279,9 +279,9 @@ mod spec {
     }
 
     #[test]
-    fn errors_on_u16_overflowed_version() {
+    fn errors_on_u16_overflowed_revision() {
         let input = parse_quote! {
-            #[event(name = "event", version = 4294967295)]
+            #[event(name = "event", revision = 4294967295)]
             struct Event;
         };
 
@@ -293,7 +293,7 @@ mod spec {
     #[test]
     fn errors_on_enum() {
         let input = parse_quote! {
-            #[event(name = "event", version = 1)]
+            #[event(name = "event", revision = 1)]
             enum Event {
                 Event1(Event1),
             }