+++ src/es/event.rs
@@ -2,12 +2,12 @@
 
 #[doc(inline)]
 pub use arcane_core::es::event::{
-    Event, Initial, Initialized, Name, Sourced, Sourcing, Version, Versioned,
+    Event, Initial, Initialized, Name, Revised, Revision, Sourced, Sourcing,
 };
 
 #[cfg(feature = "derive")]
 #[doc(inline)]
-pub use arcane_codegen::es::event::{Event, Versioned};
+pub use arcane_codegen::es::event::{Event, Revised};
 #[cfg(feature = "derive")]
 #[doc(inline)]
 pub use arcane_core::es::event::codegen;